Move and rename flex_tone
