+++ example/lib/example5_themes_playground/widgets/panels/seeded_color_scheme_settings/flex_tone_config_popup_menu.dart
@@ -1,7 +1,7 @@
 import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../shared/const/flex_tone.dart';
+import '../../../../shared/model/flex_tones_enum.dart';
 import '../../../../shared/widgets/universal/list_tile_reveal.dart';
 import '../../shared/color_scheme_box.dart';
