+++ example/lib/example5_themes_playground/widgets/panels/seeded_color_scheme_settings/seeded_color_scheme_settings.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 import '../../../../shared/const/app_color.dart';
-import '../../../../shared/const/flex_tone.dart';
 import '../../../../shared/controllers/theme_controller.dart';
+import '../../../../shared/model/flex_tones_enum.dart';
 import '../../../../shared/widgets/universal/list_tile_reveal.dart';
 import '../../../../shared/widgets/universal/switch_list_tile_reveal.dart';
 import '../../shared/show_input_colors_switch.dart';
