+++ example/lib/example5_themes_playground/widgets/panels/seeded_color_scheme_settings/show_tonal_palette.dart
@@ -2,8 +2,8 @@ import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../shared/const/app_color.dart';
-import '../../../../shared/const/flex_tone.dart';
 import '../../../../shared/controllers/theme_controller.dart';
+import '../../../../shared/model/flex_tones_enum.dart';
 import '../../../../shared/widgets/universal/list_tile_reveal.dart';
 import 'select_palette_type.dart';
 import 'tonal_palette_colors.dart';
